requirements.txt
Flask>=2.0
boto3>=1.20
Flask-SQLAlchemy>=3.0
python-dateutil>=2.8
